Soften "optimized reply" language per Andrew Sullivan.
Andrew once said, “I'm a little uncomfortable with "optimized reply" as
the name
for this.  It seems to me that one could be getting this special
authoritative reply for lots of reasons, though "optimization" is the
usual one.  I'd suggest a more neutral term; perhaps "origin-tailored
reply" or something.  The point is that the response you got is
somehow scoped to some query originators and not others, and the
document needn't take a position on whether that's good or bad.”